Create README.md
CAD systems are capable of rapid and thorough analyses of designs, whether it is a simple shelf bracket or a gear in the large and complex structure. CATIA(Computer-Aided Three-Dimensional  Interactive Application) is one of the leading modelling software in aircrafts industries. Boeing 777 was the first aircraft which was completely designed by computers and like other aircrafts mockups, no prototypes and models were built. Boeing 777 was directly constructed and assembled from CAD and CAM software model that has been developed. Through CAD, the performance of structure subjected to static or fluctuating load or temperature gradients can be simulated and analyzed and also they can be tested rapidly and accurately. 
If you make a model of complete aircraft from nut and bolt to aircraft skin or surface which will be further divided into sub-parts such as wings, engine, nacelle, fuselage, vertical and horizontal stabilizer and so more. These sub-parts contains elements which have standard sizes such as nuts and bolt. Now, whenever there is need of nut and bolt, you might need to design it each and every time but using UI Automation this task can be made easy and it will save a lot of time of employees. Basically, if you see modelling more closely, it consists of some repetitive clicks and some dimensions input from the keyboard. 
If we want to make a model of let say C- clamp. The first step will be creating a new part(new file) and saving it with some XYZ name. Then you choose the sketch environment and make some sketch of it with some specific dimensions and further it will be followed by coming back to part design environment and extrude it to some pre-defined value. On an average creating model of C-clamp, a human will take 3-5 mins. But if it will be done by UI automation it will hardly take a minute.  G1ANT provide UI automation with its own very understandable and easy to learn language.  G1ANT provide robotic software solutions to businesses that allow employees to focus on higher-value activities and free them from non-creative and repeatable processes with the help of Robotic Process Automation (RPA).  Research says doing repetitive and monotonous work decreases interest and efficiency of an employee. By including UI addon in our sketch we can easily design our model. Instead of wasting our time on creating the same thing, again and again, we can save our precious time and let the UI Automation do it for us.